+++ client.go
@@ -3,6 +3,7 @@ package resolvermt
 // Client is used to send DNS requests to resolvers concurrently.
 type Client interface {
 	Resolve(domains []string, rrtype RRtype) []Record
+	QueryCount() int
 	Close()
 }